Implemented LocalPreferenceImpl as the implementation of interface LocalPreference.
